+++ bot/ui/helpers.py
@@ -4,7 +4,7 @@ import traceback
 
 import discord
 
-from shared.logs import logger
+from shared.logs import highlight, logger
 from shared.types import User
 
 
@@ -25,12 +25,12 @@ async def handle_error(error: Exception, interaction: discord.Interaction) -> No
     logger.critical(
         f"An unexpected error occurred for user '{interaction.user.display_name}' "
         f"({interaction.user.id}): "
-        f"\n{traceback.format_exc()}"
+        f"{highlight(str(error))}.\n\n{traceback.format_exc()}"
     )
 
     embed: discord.Embed = discord.Embed(color=discord.Color.red())
     embed.title = "Error"
-    embed.description = f"An unexpected error occurred: '{error}'. \nTry again later."
+    embed.description = f"An unexpected error occurred: '{error}'. \n\nTry again later."
 
     icon: discord.File = discord.File("./src/shared/images/error.png", filename="error.png")
     embed.set_thumbnail(url="attachment://error.png")